move constant since scripts are running in isolation
