+++ scripts/helpers.js
@@ -1,3 +1,5 @@
+export const REQUEST_SLEEP = 5000;
+
 export function sleep(ms = 0, msMax = null) {
   return new Promise(r =>
     setTimeout(r, msMax ? Math.floor(Math.random() * (msMax - ms)) + ms : ms)
